Fix multiple alpha divs being created
Setting event listener for mouseover seems to fire the event more than
once, possibly due to multiple divs taking up the same space.
This occasionally creates multiple alpha divs per box div, which leads
to the divs overflowing the container. This might happen due to if-else
loop firing again on the second event before the script from the first
event has actually had time to run and make the conditions for the
else-if statement false. Using mouseenter instead of mouseover seems to
fix this.